requirements.txt
albumentations
scipy
numpy
pandas
pycocotools
pyyaml
scikit-learn
scikit-multilearn
wandb
tabulate
